Byte Order Mark for Java API
