+++ csv/src/main/java/akka/stream/alpakka/csv/javadsl/CsvQuotingStyle.java
@@ -3,7 +3,13 @@
  */
 package akka.stream.alpakka.csv.javadsl;
 
+/**
+ * Select which fields to quote in CSV formatting.
+ */
 public enum CsvQuotingStyle {
+    /** Quote all fields */
     ALWAYS,
+
+    /** Quote only fields requiring quotes */
     REQUIRED;
 }
