+++ csv/src/main/scala/akka/stream/alpakka/csv/CsvParser.scala
@@ -5,6 +5,7 @@ package akka.stream.alpakka.csv
 
 import java.nio.charset.UnsupportedCharsetException
 
+import akka.stream.alpakka.csv.scaladsl.ByteOrderMark
 import akka.util.{ByteString, ByteStringBuilder}
 
 /**
