+++ csv/src/main/scala/akka/stream/alpakka/csv/scaladsl/ByteOrderMark.scala
@@ -1,7 +1,7 @@
 /*
  * Copyright (C) 2016-2017 Lightbend Inc. <http://www.lightbend.com>
  */
-package akka.stream.alpakka.csv
+package akka.stream.alpakka.csv.scaladsl
 
 import akka.util.ByteString
